+++ fdb-record-layer-lucene/src/main/java/com/apple/foundationdb/record/lucene/directory/LuceneSerializer.java
@@ -57,12 +57,15 @@ public class LuceneSerializer {
     private final boolean encryptionEnabled;
     @Nullable
     private final SerializationKeyManager keyManager;
+    private final boolean fieldProtobufPrefixEnabled;
 
     public LuceneSerializer(boolean compressionEnabled,
-                            boolean encryptionEnabled, @Nullable SerializationKeyManager keyManager) {
+                            boolean encryptionEnabled, @Nullable SerializationKeyManager keyManager,
+                            boolean fieldProtobufPrefixEnabled) {
         this.compressionEnabled = compressionEnabled;
         this.encryptionEnabled = encryptionEnabled;
         this.keyManager = keyManager;
+        this.fieldProtobufPrefixEnabled = fieldProtobufPrefixEnabled;
     }
 
     public boolean isCompressionEnabled() {
@@ -71,6 +74,10 @@ public class LuceneSerializer {
 
     public boolean isEncryptionEnabled() {
         return encryptionEnabled;
+    }
+
+    public boolean isFieldProtobufPrefixEnabled() {
+        return fieldProtobufPrefixEnabled;
     }
 
     @Nullable
@@ -131,7 +138,7 @@ public class LuceneSerializer {
             return null;
         }
 
-        if (data.length < 2) {
+        if (data.length < 1) {
             throw new RecordCoreException("Invalid data")
                     .addLogInfo(LuceneLogMessageKeys.DATA_VALUE, data);
         }
@@ -269,5 +276,47 @@ public class LuceneSerializer {
             CipherPool.returnCipher(cipher);
         }
         encodedDataInput.reset(decrypted);
+    }
+
+    @Nullable
+    public byte[] encodeFieldProtobuf(@Nullable byte[] bytes) {
+        if (fieldProtobufPrefixEnabled) {
+            return encode(bytes);
+        } else {
+            return bytes;
+        }
+    }
+
+    @Nullable
+    public byte[] decodeFieldProtobuf(@Nullable byte[] bytes) {
+        if (bytes == null) {
+            return null;
+        }
+
+        if (isProtobufMessageWithoutPrefix(bytes)) {
+            return bytes;
+        }
+        return decode(bytes);
+    }
+
+    // This can be removed once it is guaranteed that all indexes are using the encoded format.
+    // Only works for Protobuf messages all of whose fields are themselves length-delimited,
+    // such as LuceneStoredFields (StoredField or bytes) or FieldInfos (FieldInfo).
+    private boolean isProtobufMessageWithoutPrefix(@Nonnull byte[] bytes) {
+        if (bytes.length < 1) {
+            return true;    // No room for prefix; empty message.
+        }
+        final int byte0 = bytes[0];
+        final int fieldTypeOrPrefixFlags = byte0 & 7;
+        // Either Protobuf LEN or ENCODING_COMPRESSED.
+        if (fieldTypeOrPrefixFlags != 2) {
+            return false;
+        }
+        final int fieldNumberOrKeyNumber = byte0 >> 3;
+        // ENCODING_COMPRESSED will never have a key; 0 is not a valid field number.
+        if (fieldNumberOrKeyNumber == 0) {
+            return false;
+        }
+        return true;
     }
 }