+++ fdb-relational-core/src/test/java/com/apple/foundationdb/relational/recordlayer/query/ExplainTests.java
@@ -135,7 +135,7 @@ public class ExplainTests {
                         final var continuationInfo = resultSet.getStruct(5);
                         org.junit.jupiter.api.Assertions.assertNotNull(continuationInfo);
                         final var assertStruct = RelationalStructAssert.assertThat(continuationInfo);
-                        assertStruct.hasValue("EXECUTION_STATE", new byte[]{0, 21, 1, 21, 11});
+                        assertStruct.hasValue("EXECUTION_STATE", new byte[]{10, 5, 0, 21, 1, 21, 11, 17, -84, -51, 115, -104, -35, 66, 0, 94});
                         assertStruct.hasValue("VERSION", 1);
                         assertStruct.hasValue("PLAN_HASH_MODE", "VC0");
                         assertStruct.hasValue("PLAN_HASH", -1635569052);
@@ -167,7 +167,7 @@ public class ExplainTests {
                         final var continuationInfo = resultSet.getStruct(5);
                         org.junit.jupiter.api.Assertions.assertNotNull(continuationInfo);
                         final var assertStruct = RelationalStructAssert.assertThat(continuationInfo);
-                        assertStruct.hasValue("EXECUTION_STATE", new byte[]{0, 21, 1, 21, 11});
+                        assertStruct.hasValue("EXECUTION_STATE", new byte[]{10, 5, 0, 21, 1, 21, 11, 17, -84, -51, 115, -104, -35, 66, 0, 94});
                         assertStruct.hasValue("VERSION", 1);
                         assertStruct.hasValue("PLAN_HASH_MODE", "VC0");
                         assertStruct.hasValue("PLAN_HASH", -1635569052);
@@ -199,7 +199,7 @@ public class ExplainTests {
                         final var continuationInfo = resultSet.getStruct(5);
                         org.junit.jupiter.api.Assertions.assertNotNull(continuationInfo);
                         final var assertStruct = RelationalStructAssert.assertThat(continuationInfo);
-                        assertStruct.hasValue("EXECUTION_STATE", new byte[]{0, 21, 1, 21, 11});
+                        assertStruct.hasValue("EXECUTION_STATE", new byte[]{10, 5, 0, 21, 1, 21, 11, 17, -84, -51, 115, -104, -35, 66, 0, 94});
                         assertStruct.hasValue("VERSION", 1);
                         assertStruct.hasValue("PLAN_HASH_MODE", "VC0");
                         assertStruct.hasValue("PLAN_HASH", -1635569052);